refactor: home filter component with sprint-3 UI
